pyproject.toml
[project]
name = "tpdw-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.0",
    "python-telegram-bot>=22.0",
    "redis==6.0.0b2",
]
